requirements.txt
# MUKAprint Dependencies

# GUI Framework
PySide6>=6.4.0

# File System Monitoring
watchdog>=2.1.9

# Windows Integration
pywin32>=305

# Document Processing
python-docx>=0.8.11
PyPDF2>=2.11.1
openpyxl>=3.0.10

# Utilities
pillow>=9.3.0  # For image handling
qtawesome>=1.2.1  # For icons in the UI